pyproject.toml
[tool.poetry]
name = "ultimate_sprint_report"
version = "0.1.0"
description = "A library for generating sprint reports from Jira data."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/ultimate_sprint_report"
repository = "https://github.com/yourusername/ultimate_sprint_report"
keywords = ["jira", "sprint", "report", "analytics"]

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.4"
numpy = "^1.21"
pandas = "^1.3"
atlassian-python-api = "^3.21"
tqdm = "^4.62"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"